dialer: rename package from utils to dialer, standardize key name
