+++ dialer/context.go
@@ -1,7 +1,9 @@
-package utils
+package dialer
 
 import (
 	"context"
+
+	"github.com/RTradeLtd/grpc/middleware"
 
 	"google.golang.org/grpc/metadata"
 )
@@ -9,6 +11,6 @@ import (
 // SecureRequestContext attaches given key as metadata to the provided context
 func SecureRequestContext(ctx context.Context, key string) context.Context {
 	return metadata.NewOutgoingContext(ctx, metadata.Pairs(
-		"key", key,
+		middleware.AuthorizationKey, key,
 	))
 }
